+++ app/src/main/java/org/simple/clinic/feature/Feature.kt
@@ -25,5 +25,6 @@ enum class Feature(
   PatientReassignment(false, "patient_reassignment_v0"),
   PatientStatinNudge(false, "patient_statin_nudge_v0"),
   NonLabBasedStatinNudge(false, "non_lab_based_statin_nudge"),
-  LabBasedStatinNudge(false, "lab_based_statin_nudge")
+  LabBasedStatinNudge(false, "lab_based_statin_nudge"),
+  Screening(false, "screening_feature_v0")
 }