+++ app/src/main/java/org/simple/clinic/summary/medicalhistory/ui/MedicalHistorySummary.kt
@@ -28,6 +28,7 @@ fun MedicalHistorySummary(
     diabetesManagementEnabled: Boolean,
     showSmokerQuestion: Boolean,
     showSmokelessTobaccoQuestion: Boolean,
+    isScreeningFeatureEnabled: Boolean,
     modifier: Modifier = Modifier,
     onAnswerChange: (MedicalHistoryQuestion, Answer) -> Unit,
 ) {
@@ -42,6 +43,7 @@ fun MedicalHistorySummary(
         hypertensionAnswer = hypertensionAnswer,
         diabetesAnswer = diabetesAnswer,
         showDiabetesDiagnosisView = diabetesManagementEnabled,
+        isScreeningFeatureEnabled = isScreeningFeatureEnabled,
         onAnswerChange = onAnswerChange,
     )
 
@@ -80,6 +82,7 @@ private fun MedicalHistorySummaryPreview() {
         diabetesManagementEnabled = true,
         showSmokerQuestion = false,
         showSmokelessTobaccoQuestion = false,
+        isScreeningFeatureEnabled = true,
         onAnswerChange = { _, _ ->
           // no-op
         }
@@ -102,6 +105,7 @@ private fun MedicalHistorySummaryNoDiabetesManagementPreview() {
         diabetesManagementEnabled = false,
         showSmokerQuestion = false,
         showSmokelessTobaccoQuestion = false,
+        isScreeningFeatureEnabled = true,
         onAnswerChange = { _, _ ->
           // no-op
         }
@@ -124,6 +128,7 @@ private fun MedicalHistorySummarySmokerPreview() {
         diabetesManagementEnabled = true,
         showSmokerQuestion = true,
         showSmokelessTobaccoQuestion = false,
+        isScreeningFeatureEnabled = true,
         onAnswerChange = { _, _ ->
           // no-op
         }
@@ -146,6 +151,7 @@ private fun MedicalHistorySummaryTobaccoUsePreview() {
         diabetesManagementEnabled = true,
         showSmokerQuestion = true,
         showSmokelessTobaccoQuestion = true,
+        isScreeningFeatureEnabled = true,
         onAnswerChange = { _, _ ->
           // no-op
         }