+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 
 - Hide next button when registration is going
 - Add Suspected option to hypertension and diabetes questions
+- Handle Screening feature visibility based on feature flag `Screening`
 
 ## 2025.09.09